requirements.txt
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
backports.zoneinfo==0.2.1
Django==3.2.25
django-filter==23.5
django-redis==5.4.0
django_debug_toolbar==3.8.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.0
drf-spectacular==0.27.2
geographiclib==2.0
geopy==2.4.1
importlib-metadata==6.7.0
importlib-resources==5.12.0
inflection==0.5.1
jsonschema==4.17.3
Markdown==3.4.4
numpy==1.21.6
pkgutil_resolve_name==1.3.10
psycopg2-binary==2.9.9
PyJWT==2.8.0
pyrsistent==0.19.3
python-decouple==3.8
pytz==2024.1
PyYAML==6.0.1
redis==5.0.4
sqlparse==0.4.4
typing_extensions==4.7.1
uritemplate==4.1.1
zipp==3.15.0
